+++ CHANGELOG.rst
@@ -5,6 +5,7 @@
 * update enums.py [tony612]
 * add gen-enums.py [jcupitt]
 * improve custom source/target types [jcupitt]
+* revise types for set_blob [jcupitt]
 
 ## Version 2.1.11 (7 Nov 2019)