+++ CHANGELOG.rst
@@ -7,6 +7,7 @@
 * improve custom source/target types [jcupitt]
 * revise types for set_blob to fix exception with old libvips [jcupitt]
 * fix 32-bit support [dqxpb]
+* remove pytest-runner from pipy deps [lgbaldoni]
 
 ## Version 2.1.11 (7 Nov 2019)