+++ CHANGELOG.rst
@@ -1,5 +1,9 @@
 # master 
 
+## Version 2.1.13 (4 Jul 2020)
+
+* better diagnostics for API mode install failure [kleisauke]
+
 ## Version 2.1.12 (17 Feb 2020) 
 
 * update enums.py [tony612]